Add actual panic message to expect
A panic halts your problem completely, to fail fast is the goal.
Calling expect() on a Result might trigger one. It indicates not
handling the error programatically and just wanting to provide a nice
message.
However if you provide an empty message, you might just call `unwrap()`
and get the actual error.